Extract conversation processor utility funcs from gpt.py into utils.py
